modif de la requete de tous les resto par rating decroissant

+++ EatBrusselsAPI/EatBrussels/Models/RatingModel.cs
@@ -9,6 +9,7 @@ namespace EatBrussels.Models
     {
         public int restaurantID { get; set; }
         public int restaurantName { get; set; }
-        public float averageRating { get; set; }
+        public int averageRating { get; set; }
+
     }
 }
